完善温湿度系统的代码，上传bootloader代码和恢复出厂设置代码

+++ Detect_Environment/bsp/IR.c
@@ -30,7 +30,7 @@ uint8_t IR_GetRepeatFlag(void)
 void IR_EXTI_Callback(void)
 {
 	static uint8_t s=0, IRNum;
-	uint64_t a;
+	uint16_t a;
 	uint8_t i;
 	
 	if(s==0)
